add inheritance asserv_default and asservExtended. add inheritance actions_default and actionsExtended. bug using several moveD...to be fixed.

+++ trunk/trunk-fusion/src/Common/Asserv.Driver/AAsservDriver.hpp
@@ -25,8 +25,8 @@ public:
 	{
 	}
 
-	virtual void setMotorLeftPosition(int tick, int power) = 0;
-	virtual void setMotorRightPosition(int tick, int power)= 0;
+	virtual void setMotorLeftPosition(long tick, int power) = 0;
+	virtual void setMotorRightPosition(long tick, int power)= 0;
 
 	virtual void setMotorLeftPower(int power, int timems) = 0;
 	virtual void setMotorRightPower(int power, int timems)= 0;
